Fixed how elb_name is deduced from elb_dns_name
In cases where ELB is not autogenerated by Aviatrix, it is not valid to
get the 31 first characters to get elb_name from elb_dns_name.
No it is obtained removing what's after the first . (DNS zone name), and
then the last section starting with a - in the hostname, which is
introduced with a random value by AWs.